refactor: Command 클래스 위치 변경, 불필요한 메소드 제거
